Move geolocation call to map view, use geolocation as region and marker coordinates.
